+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
+- geyser: add option `channel_capacity_filters` ([#432](https://github.com/rpcpool/yellowstone-grpc/pull/432))
+
 ### Breaking
 
 ## 2024-10-04